add @Valid to methods and throw CardNotFoundException

+++ src/main/java/com/logicea/cards/CardController.java
@@ -1,5 +1,6 @@
 package com.logicea.cards;
 
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,17 +20,17 @@ public class CardController {
     }
 
     @GetMapping("/{card_id}")
-    Card one(@PathVariable int card_id) {
+    Card one(@PathVariable int card_id) throws CardNotFoundException {
         return repository.findById(card_id).orElseThrow(() -> new CardNotFoundException(card_id));
     }
 
     @PostMapping
-    Card newCard(@RequestBody Card newCard) {
+    Card newCard(@Valid @RequestBody Card newCard) {
         return repository.save(newCard);
     }
 
     @PutMapping("/{id}")
-    Card replaceCard(@RequestBody Card newCard, @PathVariable int id) {
+    Card replaceCard(@Valid @RequestBody Card newCard, @PathVariable int id) throws CardNotFoundException {
         return repository.findById(id).map(card -> {
             card.setCardId(newCard.getCardId());
             card.setName(newCard.getName());
@@ -38,13 +39,12 @@ public class CardController {
             card.setStatus(newCard.getStatus());
             card.setUserId(newCard.getUserId());
             return repository.save(newCard);
-        }).orElseGet(() -> {
-            return repository.save(newCard);
-        });
+        }).orElseThrow(() -> new CardNotFoundException(id));
     }
 
+
     @PatchMapping("/{id}")
-    Card partialUpdateCard(@RequestBody Card updates, @PathVariable int id) {
+    Card partialUpdateCard(@Valid @RequestBody Card updates, @PathVariable int id) throws CardNotFoundException {
         return repository.findById(id).map(card -> {
             if (updates.getName() != null) {
                 card.setName(updates.getName());
@@ -67,7 +67,14 @@ public class CardController {
     }
 
     @DeleteMapping("/{id}")
-    void deleteCard(@PathVariable int id) {
-        repository.deleteById(id);
+    void deleteCard(@PathVariable int id) throws CardNotFoundException {
+        if(repository.findById(id).isPresent()) {
+            repository.deleteById(id);
+        }
+        else {
+            throw new CardNotFoundException(id);
+        }
     }
+
+
 }
